requirements.txt
Django==1.9.7
django-betterforms==1.1.4
selenium==2.53.6
six==1.10.0
